chore(web): split large regime and ai panel sections
